requirements.txt
faker==19.12.1
isort==5.12.0
locust==2.14.0
playwright==1.39.0
pydantic==2.4.2
pytest==7.4.3
pytest-playwright==0.4.3
pytest-schema==0.1.1
python-dotenv==1.0.0
requests==2.28.1
allure-pytest==2.13.2
